Example Datasets for Robitzsch and Steinfeld (2018)

Example datasets for Robitzsch and Steinfeld (2018).

data(data.ptam1)
data(data.ptam2)
data(data.ptam3)
data(data.ptam4)
data(data.ptam4long)
data(data.ptam4wide)

Format

The dataset data.ptam1 is a subset of the dataset from Example 3 of the ConQuest manual and contains 9395 ratings for 6877 students and 9 raters on 2 items (OP and TF). The format is

'data.frame': 9395 obs. of 4 variables:
$ pid : int 1508 1564 1565 1566 1567 1568 1569 1629 1630 1631 ...
$ rater: num 174 124 124 124 124 124 124 114 114 114 ...
$ OP : int 2 1 2 1 1 1 2 2 2 3 ...
$ TF : int 3 1 2 2 1 1 2 2 2 3 ...

The dataset data.ptam2 contains 1043 ratings for 262 students and 17 raters on 19 items (A1, ..., D9). The format is

'data.frame': 1043 obs. of 21 variables:
$ idstud : int 1001 1001 1001 1001 1002 1002 1002 1002 1003 1003 ...
$ idrater: int 101 108 212 215 104 108 209 211 103 104 ...
$ A1 : int 1 1 1 1 1 1 1 1 1 1 ...
$ A2 : int 1 1 1 1 0 0 0 1 1 1 ...
$ A3 : int 1 1 1 1 1 1 0 1 0 0 ...
[...]
$ D9 : int 2 2 2 2 2 2 2 2 1 0 ...

The dataset data.ptam3 contains 523 ratings for 262 students and 8 raters on 23 items (A1, ..., J0). The format is

'data.frame': 523 obs. of 25 variables:
$ idstud : int 1001 1001 1002 1002 1003 1003 1004 1004 1005 1005 ...
$ idrater: int 101 108 104 108 103 104 102 104 102 108 ...
$ A1 : int 1 1 1 1 1 1 1 1 1 1 ...
$ A2 : int 1 1 0 0 1 1 NA 0 1 1 ...
$ A3 : int 1 1 1 1 0 0 0 0 0 0 ...
[...]
$ J0 : int 2 3 3 2 0 0 2 2 0 1 ...

The dataset data.ptam4 contains 592 ratings for 209 students and 10 raters on 3 items (crit2, crit3 and crit4). The format is

'data.frame': 592 obs. of 5 variables:
$ idstud: num 10005 10009 10010 10010 10014 ...
$ rater : num 802 802 844 802 837 824 820 803 816 844 ...
$ crit2 : int 3 2 0 2 1 0 2 1 1 0 ...
$ crit3 : int 3 2 1 2 2 2 2 2 2 2 ...
$ crit4 : int 2 1 2 1 2 2 2 2 2 2 ...

The dataset data.ptam4long is the dataset data.ptam4 which has been converted into a long format for analysis with mixed effects models in the lme4 package. The format is

'data.frame': 1776 obs. of 17 variables:
$ idstud : num 10005 10005 10005 10009 10009 ...
$ rater : num 802 802 802 802 802 802 844 802 844 802 ...
$ item : Factor w/ 3 levels "crit2","crit3",..: 1 2 3 1 2 3 1 1 2 2 ...
$ value : int 3 3 2 2 2 1 0 2 1 2 ...
$ I_crit2: num 1 0 0 1 0 0 1 1 0 0 ...
$ I_crit3: num 0 1 0 0 1 0 0 0 1 1 ...
$ I_crit4: num 0 0 1 0 0 1 0 0 0 0 ...
$ R_802 : num 1 1 1 1 1 1 0 1 0 1 ...
$ R_803 : num 0 0 0 0 0 0 0 0 0 0 ...
[...]
$ R_844 : num 0 0 0 0 0 0 1 0 1 0 ...

The dataset data.ptam4wide contains multiple ratings of 40 students from the dataset data.ptam4 from the item crit2. Each column corresponds to one rater. The format is

'data.frame': 40 obs. of 11 variables:
$ pid : chr "10014" "10085" "10097" "10186" ...
$ R802: int 2 3 2 2 2 1 1 2 2 2 ...
$ R803: int 1 1 3 1 2 0 0 0 1 0 ...
$ R810: int 1 2 2 2 1 0 1 1 2 1 ...
$ R816: int 1 2 3 2 2 0 1 1 2 1 ...
$ R820: int 2 2 2 2 1 1 1 1 1 1 ...
$ R824: int 0 3 2 3 2 0 0 1 2 1 ...
$ R831: int 1 2 2 2 1 0 0 0 1 1 ...
$ R835: int 0 1 2 2 1 1 0 0 2 1 ...
$ R837: int 1 2 3 2 2 0 1 1 2 2 ...
$ R844: int 0 2 3 2 2 0 0 0 1 3 ...

References

Robitzsch, A., & Steinfeld, J. (2018). Item response models for human ratings: Overview, estimation methods, and implementation in R. Psychological Test and Assessment Modeling, 60(1), 101-139.