Composite Conditional Maximum Likelihood Estimation for the Partial Credit Model with a Design Matrix for Item Parameters

Estimates the partial credit model with a design matrix for item parameters with composite conditional maximum likelihood estimation. The estimation uses pairs of items \(X_i\) and \(X_j\) and considers conditional likelihoods \(P(X_i=k, X_j=h | \theta) / P( X_i + X_j=k+h| \theta )\). By using this strategy, the trait \(\theta\) cancels out (like in conditional maximum likelihood estimation). The proposed strategy is a generalization of the Zwinderman (1995) composite conditional maximum likelihood approach of the Rasch model to the partial credit model. See Varin, Reid and Firth (2011) for a general introduction to composite conditional maximum likelihood estimation.

immer_ccml( dat, weights=NULL, irtmodel="PCM", A=NULL, b_fixed=NULL, control=NULL )

# S3 method for immer_ccml
summary(object, digits=3, file=NULL, ...)

# S3 method for immer_ccml
coef(object, ...)

# S3 method for immer_ccml
vcov(object, ...)

Arguments

dat
Data frame with polytomous item responses \(0,1,\ldots, K\)

weights
Optional vector of sampling weights

irtmodel
Model string for specifying the item response model

A
Design matrix (items \(\times\) categories \(\times\) basis parameters). Entries for categories are for \(1,\ldots,K\)

b_fixed
Matrix with fixed \(b\) parameters

control
Control arguments for optimization function stats::nlminb

object
Object of class immer_ccml

digits
Number of digits after decimal to print

file
Name of a file in which the output should be sunk

…
Further arguments to be passed.

Details

The function estimates the partial credit model as \(P(X_i=h | \theta ) \propto \exp( h \theta - b_{ih} )\) with \(b_{ih}=\sum_l a_{ihl} \xi_l\) where the values \(a_{ihl}\) are included in the design matrix A and \(\xi_l\) denotes basis item parameters.

Value

List with following entries (selection)

coef

Item parameters

vcov

Covariance matrix for item parameters

se

Standard errors for item parameters

nlminb_result

Output from optimization with stats::nlminb

suff_stat

Used sufficient statistics

ic

Information criteria

References

Varin, C., Reid, N., & Firth, D. (2011). An overview of composite likelihood methods. Statistica Sinica, 21, 5-42.

Zwinderman, A. H. (1995). Pairwise parameter estimation in Rasch models. Applied Psychological Measurement, 19(4), 369-375.

See also

Examples

# NOT RUN {
#############################################################################
# EXAMPLE 1: Partial credit model with CCML estimation
#############################################################################

library(TAM)

data(data.gpcm, package="TAM")
dat <- data.gpcm

#-- initial MML estimation in TAM to create a design matrix
mod1a <- TAM::tam.mml(dat, irtmodel="PCM2")
summary(mod1a)

#* define design matrix
A <- - mod1a$A[,-1,-1]
A <- A[,,-1]
str(A)

#-- estimate model
mod1b <- immer::immer_ccml( dat, A=A)
summary(mod1b)
# }